move error types into types.ts ; rename compile.ts to generate.ts
