+++ packages/houdini/cmd/generators/typescript.test.ts
@@ -7,7 +7,7 @@ import * as recast from 'recast'
 import * as typeScriptParser from 'recast/parsers/typescript'
 // local imports
 import '../../../../jest.setup'
-import { runPipeline } from '../compile'
+import { runPipeline } from '../generate'
 import { mockCollectedDoc } from '../testUtils'
 
 // the config to use in tests
